drivers/: mtk-vcu: Remove duplicated build rule

+++ drivers/power/supply/mediatek/charger/ln8000_charger_2.c
@@ -1563,7 +1563,7 @@ static int ln8000_get_dev_role(struct i2c_client *client)
 
         dev_info(&client->dev,"%s: matched to %s\n", __func__, of_id->compatible);
 
-        return (int)of_id->data;
+        return (long)of_id->data;
 }
 
 static int ln8000_init_irq(struct ln8000_info *info)
